Minor corrections, as discussed.
 * MolochTileEntity will process the `actionQueue` once, making a new
queue to prevent the potential unending actions.
 * `interval` mimimum time was forced to one second.
 * Updated the Readme to reflect changes to MolochScreen Item errors,
further specify details of how variance affects interval for the
non-stats folks and indicated the minimum overriding time for interval.
 * Updated MolochScreen to leave out the badges for now (extracted
function, commented it) and moved the flame-drawing further up as the
badge drawing broke it, and we don't want that happening when bacdges
comes back.
 * Fixed the default Command to actually `/say`, though I don't know if
that was strictly needed... but I like my commands to look like
commands.